added tests for db_connector.
