--- tests/test_dbConnector.py
@@ -1,17 +1,0 @@
-from unittest import TestCase
-from db_connector import DbConnector
-
-
-class TestDbConnector(TestCase):
-    def test_open_connection(self):
-        connector = DbConnector()
-        self.fail()
-
-    def test_close_connection(self):
-        self.fail()
-
-    def test_get_connection(self):
-        self.fail()
-
-    def test_send_query(self):
-        self.fail()
